requirements.txt
fastapi
uvicorn[standard]
numpy
pandas
lightgbm
boto3
Pillow
psycopg2
python-jose
passlib
